% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_plot.R
\name{calibration_plot}
\alias{calibration_plot}
\title{generates the calibration plot for the given data.}
\usage{
calibration_plot(data, time = 5, n_bins = 10, xlim = 1, ylim = 1)
}
\arguments{
\item{data}{`tibble` or `data.frame` that contains the predicted values in a variable called `pred` and the survival object in a variable called `surv`. The survival outcome object must be created with the package `survival`}
}
\value{
a calibration plot for the given data
}
\description{
generates the calibration plot for the given data.
}
