% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrell_c_index.R
\name{harrell_c_index}
\alias{harrell_c_index}
\title{this function is a wrapper for the `Hmisc::rcorr.cens` function of the `Hmisc`
package.}
\usage{
harrell_c_index(data)
}
\arguments{
\item{data}{`tibble` or `data.frame` that contains the predicted values in a variable called `pred` and the survival object in a variable called `surv`. The survival outcome object must be created with the package `survival`}
}
\value{
the Harrell's c-index for the given data
}
\description{
this function is a wrapper for the `Hmisc::rcorr.cens` function of the `Hmisc`
package.
}
\examples{
set.seed(123)
for (i in 1:2) {
  n1 <- c(24, 138)[i]
  n0 <- c(62, 102)[i]
  y <- c(rep(0, n0), rep(1, n1))
  times <- ifelse(y == 1, rnorm(2, 1), rnorm(4,1))
  y <- survival::Surv(time = times, event = y)
  d <- 3
  x <- c(rnorm(n0, 0), rnorm(n1, d))
  harrell_c_index(tibble::tibble(pred = x, surv = y))
}
}
