% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{predicts the cardim model predictions for the given data}
\usage{
predict(data)
}
\arguments{
\item{data}{`data.frame` or `tibble` where all the patients data is stored}
}
\value{
cardim predictions
}
\description{
predicts the cardim model predictions for the given data
}
\examples{

predict(
  tibble::tibble(
    id = 1,
    age = 50,
    sex = "Male"
  )
)
}
