% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardim-package.R
\docType{package}
\name{cardim-package}
\alias{cardim}
\alias{cardim-package}
\title{cardim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
