% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{calculates the cardim model predictions}
\usage{
predict(data)
}
\arguments{
\item{data}{`data.frame` or `tibble` where all the patients data is stored}
}
\value{
cardim predictions for the given cohort
}
\description{
Calculates the cardim model predictions by using the [predict_males()] and [predict_females()] functions.
The `data.frame` must have the following structure.

| Variable | Type |
|--------|-----|
| Id | numeric or string  |
| sex | "Male" or "Female"  |
| Age, years | numeric  |
| Income less than \eqn{18000}€ | dichotomic |
| Diabetes duration, years | numeric |
| hba1c, \eqn{\%} | numeric |
| Hypertension treatment at baseline | dichotomic |
| Logarithm of albumine creatinine ratio, log(mg/g) | numeric |
| \eqn{\tqx{Total cholesterol} - \text{hdl cholesterol}}, mmol/L | numeric |
| Physical activity, inactive | dichotomic |
| Physical activity, partially active | dichotomic |
| Atrial fibrillation at baseline | dichotomic |
| \eqn{\text{Systolic blood pressure} - \text{diastolic blood pressure}}, mmHg | numeric |
| Retinopathy at baseline | dichotomic |
| Smoking status, smoker | dichotomic |
| Smoking status, ex-smoker | dichotomic |
}
\examples{

predict(
  tibble::tibble(
  id = c(1,2,3),
  sex = c("Male", "Female", "Female"),
  age = c(60, 65, 45),
  diabetes_duration = c(6, 10, 2),
  hba1c = c(9.8, 6.7, 8.5),
  hypertension_treatment = c(1, 1, 0),
  income_less_18000 = c(1, 1, 1),
  log_albtocreatratio = c(2.25, 1.84, 2.36),
  non_hdl = c(3.95, 4.3, 3.4),
  pulse_pressure = c(70, 74, 65),
  physical_activity_inactive = c(0, 0, 0),
  physical_activity_partially_active = c(1, 0, 1),
  previous_atrial_f = c(0, 0, 0),
  retinopathy = c(1, 1, 0),
  smoking_status_smoker = c(0, 0, 1),
  smoking_status_ex_smoker = c(0, 0, 0)
  )
)
}
